refactor: eliminate unnecessary wrapper classes and consolidate path utils
Issue #3: Inline single-method wrapper classes
- Inline DiffStatsManager into OutputHandler (eliminate 51-line class)
- Inline getFileDirectory into LatexDiffManager and XmlOutputManager
- Inline createContinuationMessage into ModelHandler.createContinuationPrompt
- Delete DiffStatsManager.ts, displayUtils.ts, continuationMessage.ts

Issue #5: Consolidate path normalization
- Add joinLatexPath() and ensureExtension() utilities to pathCore.ts
- Update extractFigure.ts to use shared joinLatexPath
- Update extractBibliography.ts to use shared utilities

This eliminates ~140 lines of unnecessary abstraction while improving
code locality and reducing indirection.
